Splitting up the internal server types and methods into their own package, to start packaging the components of the app
