+++ srvrepo/constraint.go
@@ -1,0 +1,12 @@
+package srvrepo
+
+// internal validation constraint definitions
+const (
+	// port range min and max values for valid server address ports.
+	portRangeMin = 1024
+	portRangeMax = 65535
+
+	// name length min and max values for valid server names.
+	nameLengthMin = 3
+	nameLengthMax = 32
+)
